+++ src/router/index.ts
@@ -1,5 +1,5 @@
 import { createRouter, createWebHistory } from "vue-router";
-import LoginSignupView from "@/views/SessionQuestionView.vue";
+import LoginSignupView from "@/views/LoginSignupView.vue";
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
